# perf(npm-resolver): reduce max memory usage