--- a/.changeset/silver-cameras-count.md
+++ b/.changeset/silver-cameras-count.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/npm-resolver": patch
+---
+
+Reduce max memory usage in npm-resolver