@jakebailey 2023-03-28
Oh, this breaks because the `saveMeta` promise is floating, violating the constraint that I outlind in the JSDoc comment. Let me see if I can fix it, but, I don't have a lot of hope given nothing awaits that promise so I don't know how it manages to complete in time for the test...