@jakebailey 2023-03-28
Alright, I got it passing; I just had to ensure that the floating promise its own handle on the `limit` entry too. The memory reduction still happens even after the update, so that's still good news!